+++ src/scenarios/simple_vm.py
@@ -10,8 +10,8 @@ from dataclasses import dataclass
 from actions import (
     TofuApplyAction,
     TofuDestroyAction,
-    StartVMAction,
-    WaitForGuestAgentAction,
+    StartProvisionedVMsAction,
+    WaitForProvisionedVMsAction,
     WaitForSSHAction,
 )
 from common import ActionResult, run_ssh
@@ -99,22 +99,19 @@ class SimpleVMConstructor:
 
             ('provision', TofuApplyAction(
                 name='provision-vm',
-                env_path='envs/test',
-            ), 'Provision test VM'),
+                env_name='test',
+            ), 'Provision VM(s)'),
 
-            ('start', StartVMAction(
-                name='start-vm',
-                vm_id_attr='test_vm_id',
+            ('start', StartProvisionedVMsAction(
+                name='start-vms',
                 pve_host_attr='ssh_host',
-            ), 'Start test VM'),
+            ), 'Start VM(s)'),
 
-            ('wait_ip', WaitForGuestAgentAction(
-                name='wait-for-ip',
-                vm_id_attr='test_vm_id',
+            ('wait_ip', WaitForProvisionedVMsAction(
+                name='wait-for-ips',
                 pve_host_attr='ssh_host',
-                ip_context_key='vm_ip',
                 timeout=180,
-            ), 'Wait for VM IP'),
+            ), 'Wait for VM IP(s)'),
 
             ('verify_ssh', WaitForSSHAction(
                 name='verify-ssh',
@@ -136,7 +133,7 @@ class SimpleVMDestructor:
         return [
             ('destroy', TofuDestroyAction(
                 name='destroy-vm',
-                env_path='envs/test',
+                env_name='test',
             ), 'Destroy test VM'),
         ]
 
@@ -157,22 +154,19 @@ class SimpleVMRoundtrip:
 
             ('provision', TofuApplyAction(
                 name='provision-vm',
-                env_path='envs/test',
-            ), 'Provision test VM'),
+                env_name='test',
+            ), 'Provision VM(s)'),
 
-            ('start', StartVMAction(
-                name='start-vm',
-                vm_id_attr='test_vm_id',
+            ('start', StartProvisionedVMsAction(
+                name='start-vms',
                 pve_host_attr='ssh_host',
-            ), 'Start test VM'),
+            ), 'Start VM(s)'),
 
-            ('wait_ip', WaitForGuestAgentAction(
-                name='wait-for-ip',
-                vm_id_attr='test_vm_id',
+            ('wait_ip', WaitForProvisionedVMsAction(
+                name='wait-for-ips',
                 pve_host_attr='ssh_host',
-                ip_context_key='vm_ip',
                 timeout=180,
-            ), 'Wait for VM IP'),
+            ), 'Wait for VM IP(s)'),
 
             ('verify_ssh', WaitForSSHAction(
                 name='verify-ssh',
@@ -182,6 +176,6 @@ class SimpleVMRoundtrip:
 
             ('destroy', TofuDestroyAction(
                 name='destroy-vm',
-                env_path='envs/test',
-            ), 'Destroy test VM'),
+                env_name='test',
+            ), 'Destroy VM(s)'),
         ]